500224478_FaceExpressionClassifier/requirements.txt
streamlit
fastapi
uvicorn[standard]
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
pillow
numpy